Split blocks.py into a submodule
